# FQDN: incrementally update selectors on DNS message