… pkg/policy/selectorcache_test.go