(Note: This is based on #29691, but does not actually rely on that PR. It's based on that just for a meaningful performance comparison. Only the last commit is new.)

This change removes a full recalculcation of FQDN policy when a DNS response is received.
    
When a response is received, one of two things can occur:
1. One or more selectors select additional IPs / identities
2. Nothing has changed
    
Not possible is the case that selectors **lose** IP addresses; that happens only upon gabage collection / entry expiration. This means we can skip a full recalculation of selectors in the critical path.
    
This change is relatively small; it adds an incremental update method to the SelectorCache, and then changes the NameManager to utilize that. It does a bit of bookkeeping to compute the delta, and passes those directly to the selectors.
    
This results in a 70% improvement in benchmark speed, as well as a 99% reduction in memory usage.

```
                     │  0_main.out  │            1_pr-29691.out            │          2_incremental.out          │
                     │    sec/op    │    sec/op     vs base                │   sec/op     vs base                │
UpdateGenerateDNS-12   3558.2m ± 1%   2157.2m ± 1%  -39.37% (p=0.000 n=10)   606.4m ± 1%  -82.96% (p=0.000 n=10)

                     │   0_main.out   │             1_pr-29691.out             │          2_incremental.out           │
                     │      B/op      │      B/op       vs base                │     B/op      vs base                │
UpdateGenerateDNS-12   902.866Mi ± 0%   532.115Mi ± 0%  -41.06% (p=0.000 n=10)   3.914Mi ± 0%  -99.57% (p=0.000 n=10)

                     │  0_main.out   │            1_pr-29691.out             │          2_incremental.out          │
                     │   allocs/op   │   allocs/op    vs base                │  allocs/op   vs base                │
UpdateGenerateDNS-12   9033.94k ± 0%   1008.96k ± 0%  -88.83% (p=0.000 n=10)   49.02k ± 0%  -99.46% (p=0.000 n=10)
```


```release-note
The latency when updating the FQDN policy subsystem in response to a DNS packet has been significantly reduced.
```