@squeed 2023-12-11
Marking as needs-discussion while we decide if this should wait for v1.16 or can go in for v1.15.